// app/lib/types.ts
export interface UIHourly {
    time: string;
    temp_c: number;
    temp_f: number;
    conditionText: string;
    conditionIcon: string;
    chance_of_rain: number;
    isNow?: boolean;
}

export interface UIDaily {
    date: string; // YYYY-MM-DD
    dayName: string; // "Mon", "Tue"
    max_c: number;
    max_f: number;
    min_c: number;
    min_f: number;
    conditionText: string;
    conditionIcon: string;
    chance_of_rain: number;
    isMock?: boolean;
}

export interface UIDetails {
    sunrise: string;
    sunset: string;
    pressure_mb: number;
    visibility_km: number;
    uv: number;
    humidity: number;
    wind_kph: number;
    feelsLike_c: number;
    feelsLike_f: number;
    airQualityIndex?: number; // US EPA Index
}

export interface UIExtraDetails {
    precip_mm: number;
    cloud: number;
    gust_kph: number;
    moon_phase: string;
    moon_illumination: number;
    us_epa_index: number;
    dewpoint_c: number;
    dewpoint_f: number;
}

export interface UIWeather {
    location: {
        name: string;
        region: string;
        country: string;
        lat: number;
        lon: number;
        localtime: string;
        tz_id: string;
    };
    current: {
        temp_c: number;
        temp_f: number;
        conditionText: string;
        conditionIcon: string;
        feelslike_c: number;
        feelslike_f: number;
        humidity: number;
        wind_kph: number;
        uv: number;
    };
    hourly: UIHourly[];
    daily: UIDaily[];
    details: UIDetails;
    extraDetails: UIExtraDetails;
}